requirements.txt
pandas==2.3.1
pillow==11.3.0
numpy==2.3.2
streamlit==1.48.1
protobuf==6.31.1
plotly==6.2.0
matplotlib==3.10.1
streamlit_option_menu==0.4.0
scipy==1.15.2
matplotlib==3.10.1
